requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==24.3.0
Brotli==1.1.0
CacheControl==0.14.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
crispy-bootstrap5==2024.10
cryptography==44.0.0
cssselect2==0.7.0
dj-database-url==2.3.0
Django==5.1.3
django-crispy-forms==2.3
django-debug-toolbar==5.0.1
django-environ==0.11.2
django-push-notifications==3.1.0
django_csp==3.8
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
et_xmlfile==2.0.0
firebase-admin==6.6.0
fonttools==4.55.3
fpdf==1.7.2
frozenlist==1.5.0
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
gunicorn==23.0.0
html5lib==1.1
http_ece==1.2.1
httplib2==0.22.0
idna==3.10
lxml==5.3.0
msgpack==1.1.0
multidict==6.1.0
mysqlclient==2.2.6
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.2
pillow==11.0.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psycopg2-binary==2.9.10
py-vapid==1.9.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.27.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.10.1
pyparsing==3.2.1
pypdf==5.1.0
pyphen==0.17.0
python-bidi==0.6.3
pywebpush==2.0.3
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
rsa==4.9
six==1.17.0
sqlparse==0.5.2
svglib==1.5.1
swapper==1.4.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.3.0
weasyprint==63.1
webencodings==0.5.1
whitenoise==6.8.2
yarl==1.18.3
zopfli==0.2.3.post1
